refactor: APIExecutor 타입 생성 후 APIRequestProtocol의 메서드 이동
